#43 Fix: Service Worker being initialized on page refresh
