+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "@storystore/pwa",
-  "version": "1.1.0",
+  "version": "1.1.1",
   "magentoDependency": ">= 2.3.5",
   "author": "Carlos A. Cabrera @fnhipster",
   "repository": "https://github.com/pmet-public/storystore-pwa.git",
